RSC convert admin working-hours page
